Feat: 장바구니-쿠폰 parallel Routes 세팅 및 디바이스 사이즈별 분기 적용
