--- src/app/(pages)/cart/@modal/(..)coupon/page.tsx
@@ -1,5 +1,0 @@
-const CouponModalInCart = () => {
-  return <div>CouponModalInCart</div>;
-};
-
-export default CouponModalInCart;
