+++ src/app/(pages)/cart/coupon/@couponModal/page.tsx
@@ -1,0 +1,10 @@
+const CouponModalInCart = () => {
+  console.log('couponModal');
+  return (
+    <div className="mt-20">
+      <div className="w-[131px] h-[100px]">d</div>
+    </div>
+  );
+};
+
+export default CouponModalInCart;
